requirements.txt
numpy
django
opencv-python
psycopg2-binary
requests
imagehash
Pillow
scikit-learn
django-cleanup

